+++ app/mortgages/new/page.tsx
@@ -1,0 +1,6 @@
+import MortgageForm from 'components/Mortgage/MortgageForm'
+import { generateNewMortage } from 'utils/mortgageCalculator'
+
+export default async function Page() {
+  return <MortgageForm mortgage={generateNewMortage()} />
+}
