+++ components/Timeline/Timeline.tsx
@@ -11,9 +11,7 @@ import styles from './Timeline.module.scss'
 import React, { useEffect, useState } from 'react'
 import Arrow, { Direction } from 'components/arrow'
 import TextNumber from 'components/textNumber'
-import Loader from 'components/loader'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import useEnsureLogin from '../../hooks/useEnsureLogin'
 import useTransactionsView from 'hooks/useTransactionsView'
 
 function TransactionsSummery({
@@ -25,7 +23,6 @@ function TransactionsSummery({
   periodResolution: TimePeriod
   maxDate: Date
 }) {
-  useEnsureLogin()
   const [isOpen, setOpen] = useState(false)
 
   return (
