--- pages/demo/index.tsx
@@ -1,10 +1,0 @@
-import { signIn } from 'next-auth/react'
-import { useEffect } from 'react'
-
-export default function DemoUser() {
-  useEffect(() => {
-    signIn('demo')
-  }, [])
-
-  return 'redirect'
-}
