--- pages/mortgages/Transactions.module.scss
@@ -1,51 +1,0 @@
-.content {
-  grid-column-start: 2;
-  grid-row-start: 3;
-}
-
-.footer {
-  grid-column-start: 2;
-  grid-row-start: 4;
-}
-
-.pageHeader {
-  display: flex;
-  align-items: center;
-  margin-bottom: .6em;
-
-  .title {
-    flex: 1;
-    font-size: 1.5em;
-  }
-}
-
-.transactionTable {
-  border-collapse: collapse;
-  width: 100%;
-  line-height: 2;
-  
-  .tableHeader {
-    background-color: var(--secondery-color);
-    color: var(--heading-font-color);
-  }
-  
-  .tableRow {
-    cursor: pointer;
-    text-align: center;
-
-    &:nth-of-type(even) {
-      background-color: var(--disabled-input-color);
-      &:hover {
-        filter: brightness(2.5);
-      }
-    }
-
-    &:nth-of-type(odd) {
-      background-color:var(--heading-font-color);
-      &:hover {
-        filter: brightness(0.9);
-      }
-    }
-  }
-}
-
